Fix subword tokenizer to handle zero hash bin size (#8093)
Closes #8051 

The hashed vocabulary table contains hash-bin coefficients and offsets. Some bins may be created empty by the hash utility. The code was not handling this case properly and resolving a hash-table index to an invalid memory location.
This PR adds a check for the hash bin size to prevent the invalid calculation.
The PR also includes a gtest with an empty hash bin entry.

Authors:
  - David Wendt (https://github.com/davidwendt)

Approvers:
  - Mark Harris (https://github.com/harrism)
  - Nghia Truong (https://github.com/ttnghia)

URL: https://github.com/rapidsai/cudf/pull/8093