Fix lists strings scatter to handle zero child rows (#8103)
Closes #8098 

The lists scatter code specialization for strings child was not checking the number of child rows equal to 0. This produces an empty strings column. This PR checks for the case and returns an empty column appropriately.
The test-case provided in the #8098 has also been incorporated in the `partition_test.cpp` gtest source.

Authors:
  - David Wendt (https://github.com/davidwendt)

Approvers:
  - Ram (Ramakrishna Prabhu) (https://github.com/rgsl888prabhu)
  - Mark Harris (https://github.com/harrism)
  - Nghia Truong (https://github.com/ttnghia)
  - Robert (Bobby) Evans (https://github.com/revans2)

URL: https://github.com/rapidsai/cudf/pull/8103