Check if a map contains a specific key (#8209)
To close https://github.com/rapidsai/cudf/issues/8120

As required in Spark 3.1.1, when ANSI mode is enabled, GetMapValue should throw an exception when the key is not found in the map in a row. 
So plugin side needs to check if a map column contains the specific key in all rows.

The new added method `mapContains` in this PR should return a column of boolean, where _false_ means key is not found.

Authors:
  - Allen Xu (https://github.com/wjxiz1992)

Approvers:
  - Jason Lowe (https://github.com/jlowe)

URL: https://github.com/rapidsai/cudf/pull/8209