Fix reductions gtests failing for debug build (#8249)
Due to an nvcc compiler bug (200529165), the standard-deviation and variance aggregations source for the `cudf::reduce` API is disabled only for a libcudf Debug build. The comments in the `std.cu` and `var.cu` indicates this bug should be fixed in 10.2 but I've verified it still occurs in 11.0, 11.1, and 11.2. The comments have been updated in this PR.

Unfortunately, the disabled source causes thet REDUCTIONS_TEST to fail for those gtests that use these two aggregation types. This PR disables those gtests when also built with Debug. This allows the gtests to complete but marked as disabled. A comment is included in the `reduction_tests.cpp` explain why the tests are disabled.

Authors:
  - David Wendt (https://github.com/davidwendt)

Approvers:
  - Vukasin Milovanovic (https://github.com/vuule)
  - Conor Hoekstra (https://github.com/codereport)

URL: https://github.com/rapidsai/cudf/pull/8249