Refactor setting stack size in regex code (#8358)
This PR is the first of several to cleanup the regex strings code. The regex code employs multiple stack sizes to perform its matching based on the number of instructions in the given regex pattern. The current implementation allocates the stack for the regex code. This PR moves this down into the regex functions themselves. This helps simplify the interface and reduce compile time a bit.

Authors:
  - David Wendt (https://github.com/davidwendt)

Approvers:
  - https://github.com/nvdbaranec
  - Devavret Makkar (https://github.com/devavret)

URL: https://github.com/rapidsai/cudf/pull/8358