Add `groupby.fillna` (#8362)
Closes #4896

This PR adds support for `groupby.fillna`. Fixes a bug in libcudf `groupby.replace_nulls` when the input is not-nullable.

Example:
```python
>>> df = cudf.DataFrame({'a': [3, 2, 3, 1], 'b': [None, 8, 8, None]}, index=['w', 'x', 'y', 'z'])
>>> df.groupby('a').fillna(42)
    b
w  42
x   8
y   8
z  42
```
Notice that the key column `a` is dropped. The index from original dataframe is kept. The row order of `fillna` columns matches that of the input. This is in contrast to a second representation from Pandas:
```python
>>> pdf = pd.DataFrame({'a': [3, 2, 3, 1], 'b': [None, 8, 8, None]}, index=['w', 'x', 'y', 'z'])
>>> pdf.groupby('a').fillna({'b': 42})
        b
a        
1 z  42.0
2 x   8.0
3 w  42.0
  y   8.0
```
Notice the rows are ordered by groups, not at their original input order. The index keeps both the keys column and the original index (reordered). Per dicsussion with @shwina, these two representations are interchangeable:
```python
>>> result = result.take(df['a'].argsort())
>>> result.index = cudf.DataFrame({'a': df['a'].sort_values(), None: result.index})
>>> result
      b
a      
1 z  42
2 x   8
3 w  42
  y   8
```
For now, in cudf we opt to support the first return format.

Authors:
  - Michael Wang (https://github.com/isVoid)

Approvers:
  - Devavret Makkar (https://github.com/devavret)
  - Ram (Ramakrishna Prabhu) (https://github.com/rgsl888prabhu)
  - Karthikeyan (https://github.com/karthikeyann)

URL: https://github.com/rapidsai/cudf/pull/8362