Rename concatenate_tests.cu to .cpp (#8555)
Found this file can be changed to .cpp by modifying the `TypedColumnTest` to build its test data in the CPU and then cudaMemcpy it to the GPU. The remain ~1000 lines did not require any updates. Only a few `lists_column_wrapper` ctors needed to be corrected as similar to PR #8527.

Authors:
  - David Wendt (https://github.com/davidwendt)

Approvers:
  - Nghia Truong (https://github.com/ttnghia)
  - https://github.com/nvdbaranec
  - Mark Harris (https://github.com/harrism)

URL: https://github.com/rapidsai/cudf/pull/8555