Remove unused cudf::strings::create_offsets (#8663)
The `create_offsets` strings API was created and ported from NVStrings for use in interchanging data with the cuDF python layer. This API no longer appears to be used or required. Removing it simplifies the `strings_column_view.hpp` which is included by many libcudf source files.

The only dependency was in some gtests files which were also updated in this PR.

The change is marked 'breaking' since a public API is being removed.

Authors:
  - David Wendt (https://github.com/davidwendt)

Approvers:
  - Vukasin Milovanovic (https://github.com/vuule)
  - Jake Hemstad (https://github.com/jrhemstad)
  - Nghia Truong (https://github.com/ttnghia)

URL: https://github.com/rapidsai/cudf/pull/8663