Use gather.hpp when gather-map exists in device memory (#9299)
This PR slightly improves libcudf compile time by changing some source files to use `gather.hpp` instead of `gather.cuh` in cases where the gather-map data exists in device-memory and therefore can be wrapped by a `column_view`. Using the `cudf::detail::gather` column-view based gather-map interface means the compiler does not have to build all of the `gather.cuh` code and kernels into the target object file. This saves build time and overall object and libcudf.so size.

Some of the changes simply removed the `gather.cuh` include in files that did not actually call `cudf::detail::gather`. The rest either used an existing `column_view` that contained the gather-map or created a `column_view` from existing memory like in a `device_uvector`.

The files that were improved are listed here.  The _rank_ value is the current ranking as reported through the ninja build log.
Baseline:
| source | size (KB) | compile time (s) | rank |
| --- | ---:| ---:| ---:|
| `src/copying/copy.cu.o` | 8579 | 219 | 1 |
| `src/groupby/hash/groupby.cu.o` | 2331 | 159 | 3 |
| `src/rolling/rolling.cu.o` <sup>1</sup> | 9865 | 235 | 5 |
| `src/lists/drop_list_duplicates.cu.o` | 2955 | 148 | 10 |
| `src/rolling/grouped_rolling.cu.o` <sup>1</sup> | 8324 | 192 | 13 |
| `src/stream_compaction/drop_nans.cu.o` <sup>2</sup> | 1284 | 51 | 71 |
| `src/stream_compaction/drop_nulls.cu.o` <sup>2</sup> | 1152 | 45 | 74 |
| `src/replace/nulls.cu.o` | 1396 | 49 | 76 |
| `src/stream_compaction/apply_boolean_mask.cu.o` <sup>2</sup> | 1365 | 47 | 80 |
| `src/dictionary/remove_keys.cu.o` | 1481 | 48 | 85 |
| `src/groupby/sort/group_replace_nulls.cu.o` | 916 | 41 | 91 |
| `src/text/generate_ngrams.cu.o` <sup>2</sup> | 1521 | 50 | 92 |
| `src/groupby/sort/group_nth_element.cu.o` | 960 | 41 | 95 |
| `src/lists/extract.cu.o` | 851 | 35 | 108 |
| `src/lists/combine/concatenate_list_elements.cu.o` | 898 | 36 | 116 |

<sup>1</sup> Changes in `src/rolling/lead_lag_nested_detail.cuh` effected these files
<sup>2</sup> Changes in `include/cudf/detail/copy_if.cuh` effected these files

Results from changes in this PR

| source | size (KB) | compile time (s) | rank |
| --- | ---:| ---:| ---:|
| `src/copying/copy.cu.o` | 7536 | 187 | 2 |
| `src/groupby/hash/groupby.cu.o` | 1236 | 124 | 17 |
| `src/rolling/rolling.cu.o` <sup>1</sup> | 8307 | 186 | 11 |
| `src/lists/drop_list_duplicates.cu.o` | 2325 | 130 | 13 |
| `src/rolling/grouped_rolling.cu.o` <sup>1</sup> | 6791 | 144 | 23 |
| `src/stream_compaction/drop_nans.cu.o` <sup>2</sup> | 674 | 35 | 96 |
| `src/stream_compaction/drop_nulls.cu.o` <sup>2</sup> | 547 | 29 | 125 |
| `src/replace/nulls.cu.o` | 773 | 33 | 100 |
| `src/stream_compaction/apply_boolean_mask.cu.o` <sup>2</sup> | 760 | 31 | 120 |
| `src/dictionary/remove_keys.cu.o` | 861 | 32 | 113 |
| `src/groupby/sort/group_replace_nulls.cu.o` | 250 | 24 | 169 |
| `src/text/generate_ngrams.cu.o` <sup>2</sup> | 948 | 35 | 109 |
| `src/groupby/sort/group_nth_element.cu.o` | 303 | 25 | 156 |
| `src/lists/extract.cu.o` | 162 | 18 | 231 |
| `src/lists/combine/concatenate_list_elements.cu.o` | 355 | 21 | 178 |

The `libcudf.so` size was only reduced by ~3MB and the overall build time did not show a significant improvement.

Although the advantage is small in the PR, this change is part of a set overall improvements to compile time on some of the slowest building source files -- those with low rank numbers in the above tables.

Authors:
  - David Wendt (https://github.com/davidwendt)

Approvers:
  - Nghia Truong (https://github.com/ttnghia)
  - Jake Hemstad (https://github.com/jrhemstad)

URL: https://github.com/rapidsai/cudf/pull/9299