Fix cudf_assert in cudf::io::orc::gpu::gpuDecodeOrcColumnData (#9348)
This fixes the `cudf_assert` check in an ORC reader device function that checks an offsets value does not exceed a `size_type`. The assert expression was just coded backwards and so was being triggered for valid values.

This PR fixes the assert expression.

The assert only occurs in a debug build since it is compiled out of a release build. Also, test code for this code path was only recently added.

Authors:
  - David Wendt (https://github.com/davidwendt)

Approvers:
  - Robert Maynard (https://github.com/robertmaynard)
  - Jake Hemstad (https://github.com/jrhemstad)
  - Karthikeyan (https://github.com/karthikeyann)

URL: https://github.com/rapidsai/cudf/pull/9348