Fix memcheck error in gtest SegmentedGatherTest/GatherSliced (#9442)
Recent run of cuda-memcheck on `gtests/COPYING_TEST` uncovered an error in the `segmented_gather_list_tests.cpp` logic for `GatherSliced`. The memcheck reported a read error on the gather-map which was created with a temporary lists column that was freed before passing it to the `cudf::lists::detail::segmented_gather` function. Fixing the testcase fixed the memcheck error.

Authors:
  - David Wendt (https://github.com/davidwendt)

Approvers:
  - Nghia Truong (https://github.com/ttnghia)
  - Bradley Dice (https://github.com/bdice)

URL: https://github.com/rapidsai/cudf/pull/9442