Expose APIs to wrap CUDA or RMM allocations with a Java device buffer instance (#9538)
Closes #9531.

Publicizes the existing Java APIs to wrap an `rmm::device_buffer` or a CUDA memory allocation as a Java device buffer instance.  This can be useful when interfacing with native code that has separately allocated memory via RMM or CUDA and is passing ownership of the buffer to a cudf object.

Authors:
  - Jason Lowe (https://github.com/jlowe)

Approvers:
  - Jim Brennan (https://github.com/jbrennan333)
  - Alessandro Bellina (https://github.com/abellina)

URL: https://github.com/rapidsai/cudf/pull/9538